requirements.txt
aiocoap
web3
rpi.gpio
smbus2
